refactor: Apply both factory and command in charge feature
