+++ src/main/java/com/snsclicksystem/main/adapter/in/charge/ChargeController.java
@@ -1,5 +1,7 @@
 package com.snsclicksystem.main.adapter.in.charge;
 
+import com.snsclicksystem.main.adapter.in.charge.factory.TossPaymentChargeFactory;
+import com.snsclicksystem.main.domain.charge.dto.ChargedInfo;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -11,11 +13,8 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import com.snsclicksystem.main.adapter.in.charge.dto.RequestCharge;
 import com.snsclicksystem.main.adapter.in.charge.dto.ResponseCharge;
 import com.snsclicksystem.main.adapter.in.common.NoCreateException;
-import com.snsclicksystem.main.adapter.in.common.NoSearchException;
-import com.snsclicksystem.main.application.port.in.charge.ChargeRequestFailException;
+import com.snsclicksystem.main.domain.charge.exception.ChargeRequestFailException;
 import com.snsclicksystem.main.application.port.in.charge.ChargeUseCase;
-import com.snsclicksystem.main.domain.charge.Charge;
-import com.snsclicksystem.main.util.modelmapper.ObjectMapper;
 
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
@@ -30,7 +29,6 @@ import lombok.extern.slf4j.Slf4j;
 public class ChargeController {
 
 	private final ChargeUseCase chargeUseCase;
-	private final ObjectMapper objectMapper;
 
 	// TODO : DELETE Test Code
 	@RequestMapping("/")
@@ -45,24 +43,8 @@ public class ChargeController {
 	@ResponseBody
 	public ResponseEntity<ResponseCharge> requestTossPayment(@RequestBody @Valid RequestCharge charge) {
 		try {
-			Charge domainCharge = Charge.createRequestCharge(charge.getMemberId(), charge.getChargeType().toString(),
-					charge.getRealAmountPaid(), charge.getMySuccessUrl(), charge.getMyFailUrl());
-			
-			Charge requestCharge = chargeUseCase.requestCharge(domainCharge);
-			
-			ResponseCharge responseCharge = new ResponseCharge();
-			responseCharge.setPayType(requestCharge.getChargeType());
-			responseCharge.setAmount(requestCharge.getRealAmountPaid());
-			responseCharge.setOrderId(requestCharge.getId());
-			responseCharge.setOrderName("포인트 충전");
-			responseCharge.setCustomerName(requestCharge.getMemberEmail());
-			responseCharge.setSuccessUrl(requestCharge.getSuccessUrl());
-			responseCharge.setFailUrl(requestCharge.getFailUrl());
-			responseCharge.setFailReason(requestCharge.getFailReason());
-			responseCharge.setCancelYN(requestCharge.getIsCancel());
-			responseCharge.setCancelReason(requestCharge.getCancelReason());
-			responseCharge.setCreatedAt(requestCharge.getCreatedAt());
-
+			ChargedInfo chargedInfo = chargeUseCase.requestCharge(new TossPaymentChargeFactory(charge));
+			ResponseCharge responseCharge = new ResponseCharge(chargedInfo);
 			return new ResponseEntity<>(responseCharge, HttpStatus.OK);
 		} catch (ChargeRequestFailException e) {
             throw new NoCreateException(e.getMessage());
