+++ src/main/java/com/snsclicksystem/main/application/service/charge/ChargeService.java
@@ -1,15 +1,13 @@
 package com.snsclicksystem.main.application.service.charge;
 
 
+import com.snsclicksystem.main.domain.charge.ChargeFactory;
+import com.snsclicksystem.main.domain.charge.dto.ChargedInfo;
+import com.snsclicksystem.main.domain.member.Member;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.StringUtils;
-
-import com.snsclicksystem.main.adapter.out.persistence.charge.ChargeEntity;
-import com.snsclicksystem.main.adapter.out.persistence.member.MemberEntity;
-import com.snsclicksystem.main.application.port.in.charge.ChargeRequestFailException;
+import com.snsclicksystem.main.domain.charge.exception.ChargeRequestFailException;
 import com.snsclicksystem.main.application.port.in.charge.ChargeUseCase;
-import com.snsclicksystem.main.application.port.out.api.charge.ChargeApi;
 import com.snsclicksystem.main.application.port.out.persistence.charge.ChargeRepository;
 import com.snsclicksystem.main.application.port.out.persistence.member.MemberRepository;
 import com.snsclicksystem.main.domain.charge.Charge;
@@ -25,33 +23,29 @@ public class ChargeService implements ChargeUseCase {
 
 	private final MemberRepository memberRepository;
 	private final ChargeRepository chargeRepository;
-	private final ChargeApi chargeApi;
-	
+
 	/*
 	 * 결제 요청 service
 	 */
 	@Override
-	public Charge requestCharge(Charge charge) throws ChargeRequestFailException{
+	public ChargedInfo requestCharge(ChargeFactory chargeFactory) throws ChargeRequestFailException{
 		//member create (TODO : DELETE)
-		memberRepository.save(new MemberEntity("ddd@gmail.com"));
-		
-		// find member
-		MemberEntity member = memberRepository.findEntityById(charge.getMemberId()).get();
-		
-		// create chargeEntity 
-		ChargeEntity chargeEntity = ChargeEntity.createChargeEntity(charge,member);
-		
-		// save request charge
-		Charge savedCharge = chargeRepository.save(chargeEntity);
-		
-		// set url from return request charge
-		savedCharge.setSuccessUrl(StringUtils.hasText(charge.getSuccessUrl()) ? charge.getSuccessUrl() : chargeApi.getSuccessUrl());
-		savedCharge.setFailUrl(StringUtils.hasText(charge.getFailUrl()) ? charge.getFailUrl() : chargeApi.getFailUrl());
-		
-		//set member email
-		savedCharge.setMemberEmail(member.getEmail());
-		
-		return savedCharge;
+//		memberRepository.save(new MemberEntity("ddd@gmail.com"));
+
+
+		//사용자를 가져온다.
+		//아마 authentication 이용해서 사용자 id나 email 받은 다음 그 정보로 조회하면 될 것 같습니다.
+//		MemberEntity member = memberRepository.findEntityById(charge.getMemberId()).get();
+		Member member = Member.builder().build();
+		//충전 시 필요한 파라미터를 받아 충전 객체를 생성한다.
+		Charge charge = chargeFactory.getCharge(ChargeParameter.builder().member(member).build());
+
+		//충전을 한다.
+		ChargedInfo chargedInfo = charge.execute();
+
+
+		//충전한 정보를 저장한다.
+		return chargeRepository.saveByChargedInfo(chargedInfo);
 	}
 
 }
