+++ src/main/java/com/snsclicksystem/main/domain/member/Member.java
@@ -6,7 +6,5 @@ import lombok.Getter;
 @Getter
 @Builder
 public class Member {
-
-	
 	private String email;
 }
